// packages/plugin-syncpolaris-api/src/utils/saving/types.ts
export interface IPolarisSaving {
  prodCode: string;
  slevel: number;
  capMethod: string;
  capAcntCode: string;
  capAcntSysNo: string;
  startDate: string;
  maturityOption: string;
  rcvAcntCode: string;
  brchCode: string;
  curCode: string;
  name: string;
  name2: string;
  termLen: string;
  maturityDate: Date;
  custCode: string;
  segCode: string;
  jointOrSingle: string;
  statusCustom: string;
  statusDate: string;
  casaAcntCode: string;
  closedBy: string;
  closedDate: string;
  lastCtDate: string;
  lastDtDate: string;
}

export interface IPolarisUpdateSaving {
  operCode: string;
  acntCode: string;
  name: string;
  name2: string;
  companyCode: string;
  prodCode: string;
  prodType: string;
  brchCode: string;
  curCode: string;
  custCode: string;
  statusSys: string;
  openDateOrg: string;
  startDate: string;
  maturityDate: Date;
  tenor: number;
  capMethod: string;
  rcvAcntCode: string;
  rcvSysNo: string;
  segCode: string;
  isCorpAcnt: number;
  jointOrSingle: string;
  lastDtDate: string;
  lastCtDate: string;
  slevel: number;
  createdBy: number;
  createdDate: Date;
  createdDatetime: Date;
  createdByName: string;
  modifiedBy: number;
  modifiedDate: Date;
  modifiedDatetime: Date;
  modifiedByName: string;
  lastSeqTxn: number;
  termBasis: string;
  termLen: number;
  passbookFacility: number;
  classNo: number;
  maturityOption: string;
  readBal: number;
  readDetail: number;
  readTran: number;
  doTran: number;
  useSpclAcnt: number;
  jointOrSingleStr: string;
  repayPriority: number;
}